(refactor) Separate function to put texture on screen
