+++ src/raycast_utils.c
@@ -6,7 +6,7 @@
 /*   By: julberna <julberna@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/02 21:28:39 by julberna          #+#    #+#             */
-/*   Updated: 2024/04/09 18:53:59 by julberna         ###   ########.fr       */
+/*   Updated: 2024/04/09 19:20:45 by julberna         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -75,16 +75,9 @@ void	draw_ceiling_floor(t_game *cub)
 
 void	draw_line(t_game *cub, int i)
 {
-	int			y;
 	int			line_height;
-	double		texture_position;
-	double		texture_x;
-	double		texture_y;
-	double		wall_x;
-	double		step;
 	t_coord		end;
 	t_coord		start;
-	int	color;
 
 	line_height = (int)(HEIGHT / cub->ray.perp_wall_dist);
 	start.y = i;
@@ -96,25 +89,7 @@ void	draw_line(t_game *cub, int i)
 	if (end.x >= HEIGHT)
 		end.x = HEIGHT - 1;
 	if (cub->ray.side == NO)
-	{
-		wall_x = cub->p1.x + cub->ray.perp_wall_dist * cub->ray.dir.x;
-		wall_x -= floor(wall_x);
-		texture_x = cub->wall->width - (wall_x * (double)cub->wall->width);
-		step = (double)cub->wall->height / line_height;
-		texture_position = (start.x - HEIGHT / HORIZON + line_height / HORIZON) * step;
-		y = start.x;
-		ft_bzero(cub->buffer, HEIGHT);
-		while (y < end.x)
-		{
-			texture_y = (int)texture_position & (cub->wall->height - 1);
-			texture_position += step;
-			color = get_color(cub->wall, texture_x, texture_y);
-			cub->buffer[y] = color;
-			y++;
-		}
-		line(cub, start, end, 0x00000000, cub->buffer);
-		ft_bzero(cub->buffer, HEIGHT);
-	}
+		put_texture(cub, start, end, line_height);
 	else if (cub->ray.side == SO)
 		line(cub, start, end, 0xffe45eff, NULL);
 	else if (cub->ray.side == EA)
@@ -122,30 +97,6 @@ void	draw_line(t_game *cub, int i)
 	else if (cub->ray.side == WE)
 		line(cub, start, end, 0xf9f9f9ff, NULL);
 }
-
-int	get_color(mlx_texture_t *texture, int texture_x, int texture_y)
-{
-	int	r;
-	int	g;
-	int	b;
-	int	a;
-	int	*p;
-
-	p = (int *)texture->pixels + texture_x + texture_y * texture->width;
-	r = ((*p) & 0xFF) >> 0;
-	g = ((*p) & 0xFF00) >> 8;
-	b = ((*p) & 0xFF0000) >> 16;
-	a = ((*p) & 0xFF000000) >> 24;
-	return (r << 24 | g << 16 | b << 8 | a);
-}
-	/* code below is more performatic */
-	// int32_t	pixel;
-
-	// pixel = *((int32_t *)texture->pixels + \
-	// (size_t)(texture_y * texture->width + texture_x));
-	// pixel = ((pixel & 0xFF) << 24) | ((pixel & 0xFF00) << 8) | \
-	// ((pixel & 0xFF0000) >> 8) | ((pixel & 0xFF000000) >> 24);
-	// return (pixel);
 
 void	calculate_frames_per_second(t_game *cub)
 {
